StoryView.Count<UInt>

+++ Sources/Xcore/SwiftUI/Components/PageView/StoryView/StoryTimer.swift
@@ -11,15 +11,15 @@ import SwiftUI
 final class StoryTimer {
     private let tick = Duration.seconds(0.01)
     private var duration: Duration
-    private let cycle: Count<Int>
+    private let cycle: Count<UInt>
     private var pagesCount: Int
     private var cyclesCompleted = 0
     private var isPaused = false
     private var timerTask: Task<Void, any Error>?
-    var onCycleComplete: ((_ remainingCycles: Count<Int>) -> Void)?
+    var onCycleComplete: ((_ remainingCycles: Count<UInt>) -> Void)?
     var progress: Double
 
-    init(pagesCount: Int, duration: Duration, cycle: Count<Int>) {
+    init(pagesCount: Int, duration: Duration, cycle: Count<UInt>) {
         self.pagesCount = pagesCount
         self.duration = duration
         self.cycle = cycle
@@ -77,7 +77,7 @@ final class StoryTimer {
                     newProgress = 0
                     onCycleComplete?(.infinite)
                 case let .finite(count):
-                    let remainingCount = max(0, count - cyclesCompleted)
+                    let remainingCount = UInt(max(0, Int(count) - cyclesCompleted))
                     onCycleComplete?(.finite(remainingCount))
 
                     if remainingCount == 0 {
